debugfs.ocfs2: Silence build errors on x86_64
Silence the following build errors on x86_64.

commands.c: In function ‘dirblocks_proxy’:
commands.c:1261: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__u64’
commands.c: In function ‘walk_refcount_block’:
commands.c:2079: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
commands.c:2096: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
dump.c: In function ‘dump_inode’:
dump.c:382: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’
dump.c: In function ‘dump_extent_block’:
dump.c:505: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’
dump.c: In function ‘dump_dir_trailer’:
dump.c:596: warning: format ‘%-15lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’
dump.c:596: warning: format ‘%-15lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
dump.c:599: warning: format ‘%-15lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
dump.c: In function ‘dump_dx_root’:
dump.c:694: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’
dump.c: In function ‘dump_refcount_block’:
dump.c:1362: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’
dump.c: In function ‘dump_refcount_records’:
dump.c:1399: warning: format ‘%-20lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
stat_sysdir.c: In function ‘show_system_inode’:
stat_sysdir.c:76: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’
stat_sysdir.c:89: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’

+++ debugfs.ocfs2/stat_sysdir.c
@@ -73,7 +73,7 @@ static int show_system_inode(struct ocfs2_dir_entry *rec, uint64_t blocknr,
 		ret = traverse_extents(fs, &(di->id2.i_list), out);
 	if (ret)
 		com_err(gbls.cmd, ret, "while traversing inode at block "
-			"%"PRIu64, rec->inode);
+			"%"PRIu64, (uint64_t)rec->inode);
 
 	if (S_ISDIR(di->i_mode)) {
 		ls.fs = ls_opts->fs;
@@ -86,7 +86,7 @@ static int show_system_inode(struct ocfs2_dir_entry *rec, uint64_t blocknr,
 					(void *)&ls);
 		if (ret)
 			com_err(gbls.cmd, ret, "while iterating // at block "
-				"%"PRIu64, rec->inode);
+				"%"PRIu64, (uint64_t)rec->inode);
 		ocfs2_free(&ls.buf);
 	}
